BZ #1725250 - Mismatches for organization and location on production environment and domain
Fixes #27233 - Assign created resources to host taxonomies (#6883)

When a host is created from puppet fact import, taxable resources such
as domain and environment are created if they do not exist, but aren't
assigned to the host's taxonomies leading the host to report mismatch.
This commit changes this behaviour so that new resources will be
assigned to the same taxonomies. If a host uses an existing resource,
it does not get assigned to the host's taxonomies to prevent
unexpected modifications of existing resources.

(cherry picked from commit 27e8d26f72999bb1512c9b641430b99f58c9baf3)