BZ #1792267 - Custom reports have file extension ".text"
Fixes #28815 - keep .txt extension for text reports

If a report template does not use report_render macro, the output format
selection is disabled and it's considered a plain text. Mimetype symbol
in this case is text and we use it for suggested file extension. However
for plain text file, it's expected to be .txt. Therefore we need an
exception in the symbol to extension conversion for this case. Other
supported types work fine and we can rely on Mimetype symbols.

(cherry picked from commit 032648cc63156bbcd11df8db3086cdb7e6283599)